Make social media buttons function as links
This change makes the buttons in the bottom left corner act like normal hyperlinks. This behavior is nice to have since it allow visitors to hover over the buttons to see where they go, right click them to copy the links, and middle click them to open them in a new tab.